feat: Extracting get variables from environment function
